Restructure of connector
