+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/PSQLConfig.java
@@ -43,12 +43,12 @@ class PSQLConfig {
   /**
    * A constant that is normally used as environment variable name for the host.
    */
-  static final String PSQL_HOST = "PSQL_HOST";
+  protected static final String PSQL_HOST = "PSQL_HOST";
 
   /**
    * A constant that is normally used as environment variable name for the port.
    */
-  static final String PSQL_PORT = "PSQL_PORT";
+  protected static final String PSQL_PORT = "PSQL_PORT";
 
   /**
    * A constant that is normally used as environment variable name for the host.
@@ -63,19 +63,19 @@ class PSQLConfig {
   /**
    * A constant that is normally used as environment variable name for the user.
    */
-  static final String PSQL_USER = "PSQL_USER";
+  protected static final String PSQL_USER = "PSQL_USER";
 
   /**
    * A constant that is normally used as environment variable name for the password.
    */
-  static final String PSQL_PASSWORD = "PSQL_PASSWORD";
+  protected static final String PSQL_PASSWORD = "PSQL_PASSWORD";
 
   /**
    * A constant that is normally used as environment variable name for the schema.
    */
   private static final String PSQL_SCHEMA = "PSQL_SCHEMA";
 
-  static final String ECPS_PHRASE = "ECPS_PHRASE";
+  protected static final String ECPS_PHRASE = "ECPS_PHRASE";
 
   private final static String DEFAULT_ECPS = "default";
 
@@ -103,7 +103,7 @@ class PSQLConfig {
    * Decodes the connector parameters.
    */
   @SuppressWarnings("unchecked")
-  static Map<String, Object> decryptECPS(String ecps, String phrase) {
+  private static Map<String, Object> decryptECPS(String ecps, String phrase) {
     try {
       return new ObjectMapper().readValue(AESHelper.getInstance(phrase).decrypt(ecps), Map.class);
     } catch (Exception e) {
@@ -116,11 +116,11 @@ class PSQLConfig {
    * A method to decode connector parameters.
    */
   @SuppressWarnings("unused")
-  static String encryptCPS(String connectorParams, String phrase) throws Exception {
+  protected static String encryptCPS(String connectorParams, String phrase) throws Exception {
     return new AESHelper(phrase).encrypt(connectorParams);
   }
 
-  static String getECPS(Event event) {
+  protected static String getECPS(Event event) {
     if (event == null || event.getConnectorParams() == null || event.getConnectorParams().get("ecps") == null) {
       return DEFAULT_ECPS;
     }
@@ -133,7 +133,7 @@ class PSQLConfig {
    *
    * @return the maximal amount of concurrent connections to be used.
    */
-  int maxPostgreSQLConnections() {
+  protected int maxPostgreSQLConnections() {
     try {
       return Integer.parseInt(readEnv(PSQL_MAX_CONN), 10);
     } catch (Exception e) {
@@ -146,7 +146,7 @@ class PSQLConfig {
    *
    * @return the host of the PostgreSQL service.
    */
-  String host() {
+  protected String host() {
     final String host = readEnv(PSQL_HOST);
     if (host != null) {
       return host;
@@ -159,7 +159,7 @@ class PSQLConfig {
    *
    * @return the host of the PostgreSQL service.
    */
-  String replica() {
+  protected String replica() {
     return readEnv(PSQL_REPLICA_HOST);
   }
 
@@ -168,7 +168,7 @@ class PSQLConfig {
    *
    * @return the port of the PostgreSQL service.
    */
-  int port() {
+  protected int port() {
     final String portText = readEnv(PSQL_PORT);
     if (portText != null) {
       try {
@@ -188,7 +188,7 @@ class PSQLConfig {
    *
    * @return the database to connect to.
    */
-  String database() {
+  protected String database() {
     final String db = readEnv(PSQL_DB);
     if (db != null) {
       return db;
@@ -201,7 +201,7 @@ class PSQLConfig {
    *
    * @return the user to connect with.
    */
-  String user() {
+  protected String user() {
     final String user = readEnv(PSQL_USER);
     if (user != null) {
       return user;
@@ -214,7 +214,7 @@ class PSQLConfig {
    *
    * @return the password to connect with.
    */
-  String password() {
+  protected String password() {
     final String password = readEnv(PSQL_PASSWORD);
     if (password != null) {
       return password;
@@ -227,7 +227,7 @@ class PSQLConfig {
 
   private String applicationName;
 
-  PSQLConfig(Event event, Context context) {
+  public PSQLConfig(Event event, Context context) {
     this.context = context;
     this.connectorParams = readECPS(getECPS(event));
     this.applicationName = context.getFunctionName();
@@ -257,12 +257,12 @@ class PSQLConfig {
   }
 
 
-  boolean isReadOnly() {
+  protected boolean isReadOnly() {
     String READ_ONLY = "READ_ONLY";
     return "true".equals(readEnv(READ_ONLY));
   }
 
-  String schema() {
+  protected String schema() {
     final String schema = readEnv(PSQL_SCHEMA);
     if (schema != null) {
       return schema;
@@ -271,7 +271,7 @@ class PSQLConfig {
     return "public";
   }
 
-  String table(Event event) {
+  protected String table(Event event) {
     if (event != null && event.getSpace() != null && event.getSpace().length() > 0) {
       return event.getSpace();
     }
@@ -279,14 +279,33 @@ class PSQLConfig {
     return null;
   }
 
-  String applicationName() {
+  protected String applicationName() {
     return applicationName;
   }
+
+  protected boolean isPropertySearchActivated(){
+    if(connectorParams != null && connectorParams.get("propertySearch") == Boolean.TRUE)
+      return true;
+    return false;
+  }
+
+  protected boolean isAutoIndexingActivated(){
+    if(connectorParams != null && connectorParams.get("autoIndexing") == Boolean.TRUE)
+      return true;
+    return false;
+  }
+
+  protected Integer onDemandLimit(){
+    if(connectorParams != null && connectorParams.get("onDemandIdxLimit") != null)
+      return  (Integer) connectorParams.get("onDemandIdxLimit");
+    return null;
+  }
+
 
   public static class AESHelper {
 
     private static Map<String, AESHelper> helpers = new HashMap<>();
-    public byte[] key;
+    private byte[] key;
 
     /**
      * Returns an instance helper for this passphrase.
@@ -294,7 +313,7 @@ class PSQLConfig {
      * @param passphrase The passphrase from which to derive a key.
      */
     @SuppressWarnings("WeakerAccess")
-    public static AESHelper getInstance(String passphrase) {
+    protected static AESHelper getInstance(String passphrase) {
       if (helpers.get(passphrase) == null) {
         helpers.put(passphrase, new AESHelper(passphrase));
       }
@@ -302,7 +321,7 @@ class PSQLConfig {
     }
 
 
-    public AESHelper(String passphrase) {
+    protected AESHelper(String passphrase) {
       //noinspection UnstableApiUsage
       this.key = Arrays.copyOf(Hashing.sha256().newHasher().putBytes(passphrase.getBytes()).hash().asBytes(), 16);
     }
@@ -312,14 +331,14 @@ class PSQLConfig {
      *
      * @param data The Base 64 encoded string representation of the encrypted bytes.
      */
-    String decrypt(String data) throws IllegalBlockSizeException, BadPaddingException {
+    protected String decrypt(String data) throws IllegalBlockSizeException, BadPaddingException {
       return new String(decrypt(Base64.getDecoder().decode(data)), StandardCharsets.UTF_8);
     }
 
     /**
      * Decrypts the given bytes.
      */
-    byte[] decrypt(byte[] data) throws IllegalBlockSizeException, BadPaddingException {
+    private byte[] decrypt(byte[] data) throws IllegalBlockSizeException, BadPaddingException {
       return decryptCipher.get().doFinal(data);
     }
 
@@ -329,14 +348,14 @@ class PSQLConfig {
      * @param data The string to encode
      * @return A Base 64 encoded string, which represents the encoded bytes.
      */
-    String encrypt(String data) throws IllegalBlockSizeException, BadPaddingException {
+    protected String encrypt(String data) throws IllegalBlockSizeException, BadPaddingException {
       return new String(Base64.getEncoder().encode(encrypt(data.getBytes())));
     }
 
     /**
      * Encrypts the provided byte array.
      */
-    byte[] encrypt(byte[] data) throws IllegalBlockSizeException, BadPaddingException {
+    private byte[] encrypt(byte[] data) throws IllegalBlockSizeException, BadPaddingException {
       final Cipher ec = encryptCipher.get();
       return ec.doFinal(data);
     }
